new formating, project setup

+++ src/Path.cpp
@@ -1,6 +1,7 @@
 #include "Path.h"
 
-Path::~Path() {
+Path::~Path()
+{
     for (auto &x : this->path)
     {
         x.second.clear();
@@ -8,29 +9,27 @@ Path::~Path() {
     this->path.clear();
 }
 
-void Path::insert(Point point, PointData data) {
+void Path::insert(Point point, PointData data)
+{
     if (contains(point) == false)
     {
-        // not founded
         vector<PointData> vecData{data};
         this->path.insert({point.toString(), vecData});
     }
     else
     {
-        // founded
         this->path.at(point.toString()).push_back(data);
     }
 }
 
-void Path::insert(string key, vector<PointData> data) {
+void Path::insert(string key, vector<PointData> data)
+{
     if (contains(key) == false)
     {
-        // not founded
         this->path.insert({key, data});
     }
     else
     {
-        // founded
         for (auto x : data)
         {
             this->path.at(key).push_back(x);
@@ -38,28 +37,34 @@ void Path::insert(string key, vector<PointData> data) {
     }
 }
 
-void Path::insert(Point point, int id, int distance) {
+void Path::insert(Point point, int id, int distance)
+{
     this->insert(point, {id, distance});
 }
 
-bool Path::contains(string key) {
+bool Path::contains(string key)
+{
     auto founded = this->path.find(key);
     return (founded != this->path.end());
 }
 
-bool Path::contains(Point point) {
+bool Path::contains(Point point)
+{
     return this->contains(point.toString());
 }
 
-void Path::remove(string key) {
+void Path::remove(string key)
+{
     this->path.erase(key);
 }
 
-void Path::remove(Point point) {
+void Path::remove(Point point)
+{
     this->remove(point.toString());
 }
 
-void Path::printPathPart(PathPart part, std::ostream &stream) {
+void Path::printPathPart(PathPart part, std::ostream &stream)
+{
 
     stream << part.first << " >";
     for (auto &x : part.second)
@@ -69,14 +74,16 @@ void Path::printPathPart(PathPart part, std::ostream &stream) {
     stream << "\n";
 }
 
-void Path::printAll(std::ostream &stream) {
+void Path::printAll(std::ostream &stream)
+{
     for (auto &x : this->path)
     {
         printPathPart(x, stream);
     }
 }
 
-void Path::insertAll(vector<Point> points, int id, int distance) {
+void Path::insertAll(vector<Point> points, int id, int distance)
+{
     std::cout << "Inserting all..." << std::endl;
 
     for (auto &x : points)
@@ -87,7 +94,8 @@ void Path::insertAll(vector<Point> points, int id, int distance) {
     }
 }
 
-Path *Path::copyCollisions(int min, int max) {
+Path *Path::copyCollisions(int min, int max)
+{
     std::cout << "Searching for collisions." << std::endl;
     Path *collisions = new Path();
 
@@ -105,8 +113,9 @@ Path *Path::copyCollisions(int min, int max) {
     return collisions;
 }
 
-bool Path::validDistance(vector<PointData>& data, int min, int max) {
-    for (auto& x : data)
+bool Path::validDistance(vector<PointData> &data, int min, int max)
+{
+    for (auto &x : data)
     {
         if (!(x.getSteps() >= min && x.getSteps() <= max))
         {
@@ -117,7 +126,8 @@ bool Path::validDistance(vector<PointData>& data, int min, int max) {
     return true;
 }
 
-bool Path::validIds(vector<PointData> &data) {
+bool Path::validIds(vector<PointData> &data)
+{
     auto first = data.at(0).getId();
     for (int i = 1; i < data.size(); i++)
     {
@@ -129,15 +139,18 @@ bool Path::validIds(vector<PointData> &data) {
     return false;
 }
 
-bool Path::validPointData(vector<PointData> &data, int min, int max) {
+bool Path::validPointData(vector<PointData> &data, int min, int max)
+{
     return this->validDistance(data, min, max) && this->validIds(data);
 }
 
-unsigned long Path::size() const noexcept {
-    return this->path.size();
+unsigned long Path::size() const noexcept
+{
+    return (unsigned long)this->path.size();
 }
 
-PathPart Path::chooseFirst() {
+PathPart Path::chooseFirst()
+{
     if (this->size() > 0)
     {
         return *this->path.begin();
